fix: int형 오버플로우 수정

+++ src/main/java/domain/Lotto.java
@@ -6,33 +6,33 @@ import java.util.List;
 
 //Lotto를 관리하는 일급 컬렉션
 public class Lotto {
-    private final List<Integer> numbers;
+    private final List<Long> numbers;
 
     //automated constructor
     public Lotto() {
-        List<Integer> cardinateNumbers = new ArrayList<>();
-        for(int i=1; i<=45; i++){
+        List<Long> cardinateNumbers = new ArrayList<>();
+        for(long i=1; i<=45; i++){
             cardinateNumbers.add(i);
         }
         Collections.shuffle(cardinateNumbers);
         this.numbers = cardinateNumbers.subList(0,6);
-        this.numbers.sort(Integer::compareTo);
+        this.numbers.sort(Long::compareTo);
     }
 
     //로또 번호는 불변하므로 setter를 제공하지 않습니다.
-    public List<Integer> getNumbers(){
+    public List<Long> getNumbers(){
         return numbers;
     }
 
-    public int matchCount(WinningNumbers winningNumbers){
-        int matchCount = 0;
-        for(Integer number : numbers){
+    public long matchCount(WinningNumbers winningNumbers){
+        long matchCount = 0;
+        for(Long number : numbers){
             matchCount = countMatches(winningNumbers, number, matchCount);
         }
         return matchCount;
     }
 
-    private static int countMatches(WinningNumbers winningNumbers, Integer number, int matchCount) {
+    private static long countMatches(WinningNumbers winningNumbers, Long number, long matchCount) {
         if(winningNumbers.getWinningNumbers().contains(number)){
             matchCount++;
         }
