+++ src/main/java/domain/Money.java
@@ -1,20 +1,20 @@
 package domain;
 
 public class Money{
-    private int money;
+    private long money;
 
-    public Money(int money) throws IllegalArgumentException{
+    public Money(long money) throws IllegalArgumentException{
         if(money < 0){
             throw new IllegalArgumentException("금액은 양수여야 합니다.");
         }
         this.money = money;
     }
 
-    public int getMoney() {
+    public long getMoney() {
         return money;
     }
 
-    public void addMoney(int amount) {
+    public void addMoney(long amount) {
         this.money += amount;
     }
 }
