+++ src/main/java/view/InputView.java
@@ -6,9 +6,9 @@ import java.util.Scanner;
 public class InputView {
     static Scanner scanner = new Scanner(System.in);
 
-    public static int inputMoneyFromUser(){
+    public static long inputMoneyFromUser(){
         System.out.println("구입금액을 입력해 주세요.");
-        return scanner.nextInt();
+        return scanner.nextLong();
     }
 
     public static WinningNumbers inputWinningNumberFromUser(){
